chore: move debug tab content
