+++ ui/pages/settings-v2/debug-tab/developer-options-toggle-row-component.tsx
