+++ ui/pages/settings-v2/debug-tab/index.scss
