+++ ui/pages/settings-v2/debug-tab/sentry-test.tsx
